requirements.txt
# Core dependencies
streamlit>=1.28.0
openai>=0.28.0
requests>=2.31.0

# LangChain for conversational agent
langchain>=0.1.0
langchain-openai>=0.0.2

# Optional but recommended
python-dotenv>=1.0.0  # For environment variables
dataclasses-json>=0.6.0  # For easy serialization

# Development dependencies (optional)
pytest>=7.4.0  # For testing
black>=23.0.0  # Code formatting
flake8>=6.0.0  # Linting
mypy>=1.5.0  # Type checking